move a few directories around, run server on port 8080
