--- .eslintrc
@@ -1,23 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "extends": "airbnb",
-  "env": {
-    "es6": true,
-    "browser": true,
-    "node": true
-  },
-  "rules": {
-    "no-console": 0,
-    "eol-last": 0,
-    "no-unused-vars": 0,
-    "react/prefer-stateless-function": 0,
-    "react/jsx-one-expression-per-line": 0,
-    "react/destructuring-assignment": 0,
-    "class-methods-use-this": 0,
-    "arrow-body-style": 0,
-    "react/no-array-index-key": 0,
-    "react/jsx-tag-spacing": 0,
-    "max-len": 0,
-    "no-trailing-spaces": 0
-  }
-}
