+++ server/index.js
@@ -30,7 +30,7 @@ function normalizePort(val) {
 /**
  * Get port from environment and store in Express.
  */
-const port = normalizePort(process.env.PORT || 3000);
+const port = normalizePort(process.env.PORT || 8080);
 app.set('port', port);
 
 /**
